Decouple sprite layer to module.
