+++ public/js/layers.js
@@ -1,23 +1,3 @@
-export function createSpriteLayer(entities, width = 64, height = 64) {
-    const spriteBuffer = document.createElement('canvas');
-    spriteBuffer.width = width;
-    spriteBuffer.height = height;
-    const spriteBufferContext = spriteBuffer.getContext('2d');
-
-    return function drawSpriteLayer(context, camera) {
-        entities.forEach(entity => {
-            spriteBufferContext.clearRect(0, 0, width, height);
-
-            entity.draw(spriteBufferContext);
-
-            context.drawImage(
-                spriteBuffer,
-                entity.pos.x - camera.pos.x,
-                entity.pos.y - camera.pos.y);
-        });
-    };
-}
-
 export function createCollisionLayer(level) {
     const resolvedTiles = [];
