+++ public/js/loaders/level.js
@@ -1,7 +1,7 @@
 import {Matrix} from '../math.js';
 import Level from '../Level.js';
 import {createBackgroundLayer} from '../layers/background.js';
-import {createSpriteLayer} from '../layers.js';
+import {createSpriteLayer} from '../layers/sprites.js';
 import {loadJSON, loadSpriteSheet} from '../loaders.js';
 
 function setupCollision(levelSpec, level) {
